+++ frontend/src/pages/GamePage/components/SelectContainer/components/GameVoteStatusContainer/GameVoteStatusContainer.styled.ts
