+++ frontend/src/pages/GamePage/components/SelectContainer/components/Timer/Timer.util.ts
