+++ frontend/src/pages/GamePage/components/SelectContainer/components/Timer/hooks/timerWorker.ts
