+++ frontend/src/pages/GamePage/components/SelectContainer/components/index.ts
@@ -1,0 +1,6 @@
+import GameVoteStatusContainer from './GameVoteStatusContainer/GameVoteStatusContainer';
+import SelectButton from './SelectButton/SelectButton';
+import SelectOption from './SelectOption/SelectOption';
+import Timer from './Timer/Timer';
+
+export { GameVoteStatusContainer, SelectButton, SelectOption, Timer };
